+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 5bae4a2 - Update DAW/BAE/Unidad-2/Tarea-6
+| Author: fjrodafo - Wed, 19 Feb 2025 19:16:48 +0000
+| 
 * Commit: b5d9a87 - Update DAW/BAE/Unidad-2/Tarea-6
 | Author: fjrodafo - Wed, 19 Feb 2025 19:04:22 +0000
 |